% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/fb_init.R
\name{fb_api_most_recent_version}
\alias{fb_api_most_recent_version}
\title{Returns the most recent version of the supported Facebook Marketing API}
\usage{
fb_api_most_recent_version()
}
\value{
string
}
\description{
Returns the most recent version of the supported Facebook Marketing API
}

