% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fb_init.R
\name{fb_api_version}
\alias{fb_api_version}
\title{Returns the currently used version of the Facebook Marketing API}
\usage{
fb_api_version()
}
\value{
string
}
\description{
Returns the currently used version of the Facebook Marketing API
}
