% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/fb_insights.R
\name{fb_insights}
\alias{fb_insights}
\title{Insights}
\usage{
fb_insights(fbacc, target = fbacc$acct_path, job_type = c("sync", "async"),
  ...)
}
\arguments{
\item{fbacc}{}

\item{target}{ad account id (default), campaign id, adset id or ad id}

\item{job_type}{synchronous or asynchronous request. If the prior fails with "please reduce the amount of data", it will fall back to async request.}

\item{...}{named arguments passed to the API, like time range, fields, filtering etc.}
}
\value{
JSON
}
\description{
Insights
}
\examples{
\dontrun{
fb_insights(fbacc)
fb_insights(fbacc, date_preset = 'today', level = 'adgroup')
}
}
\references{
\url{https://developers.facebook.com/docs/marketing-api/insights/v2.3}
}

