% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/fb_reports.R
\name{fb_reportstats_ad}
\alias{fb_reportstats_ad}
\title{Get Ad Report Stats}
\usage{
fb_reportstats_ad(fbacc, ...)
}
\arguments{
\item{fbacc}{}

\item{...}{named arguments passed to the API, like time range, fields, filtering etc.}
}
\value{
list
}
\description{
Get Ad Report Stats
}
\note{
This will be soon deprecated. Use the new Insight feature, if possible.
}
\references{
\url{https://developers.facebook.com/docs/marketing-api/adreportstats/v2.3}
}

