% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/fb_reports.R
\name{fb_stats_ad}
\alias{fb_stats_ad}
\title{Get Ad Stats}
\usage{
fb_stats_ad(fbacc, ids)
}
\arguments{
\item{fbacc}{(optional) \code{FB_Ad_account} object, which defaults to the last returned object of \code{\link{fbad_init}}.}

\item{ids}{adgroup ids}
}
\value{
list (raw parsed JSON) to be post-processed
}
\description{
Get Ad Stats
}
\note{
This API endpoint is be soon deprecated and only available with v2.3. Use the new Insight feature, if possible.
}
\examples{
\dontrun{
res <- fb_stats_ad(...)
library(rlist)
list.stack(list.select(res, adgroup_id, impressions, clicks, spent, unique_impressions))
}
}
\references{
\url{https://developers.facebook.com/docs/marketing-api/adstatistics/v2.3#read}
}

