% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/fb.R
\name{fbad_add_audiance}
\alias{fbad_add_audiance}
\title{FB add people to audiance}
\usage{
fbad_add_audiance(fbacc, audiance_id, schema = c("EMAIL_SHA256",
  "PHONE_SHA256"), hashes)
}
\arguments{
\item{fbacc}{FB_Ad_account object returned by \code{fbad_init}}

\item{audiance_id}{string}

\item{schema}{only two schema are supported out of the 4}

\item{hashes}{character vector of hashes}
}
\description{
FB add people to audiance
}
\references{
https://developers.facebook.com/docs/marketing-api/custom-audience-targeting/v2.2#create
}

