% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fb_account.R
\name{fbad_assign_users_to_account}
\alias{fbad_assign_users_to_account}
\title{Assign a user to an Ad Account}
\usage{
fbad_assign_users_to_account(account_id, access_token,
  tasks = c("MANAGE", "ADVERTISE", "ANALYZE"), user)
}
\arguments{
\item{account_id}{string}

\item{tasks}{enum}

\item{user}{id}
}
\description{
Assign a user to an Ad Account
}
