% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/fb.R
\name{fbad_check_curl_params}
\alias{fbad_check_curl_params}
\title{Run basic checks on curl get/post parameters}
\usage{
fbad_check_curl_params(params)
}
\arguments{
\item{params}{named list of parameters to GET/POST}
}
\value{
list if OK, error if not
}
\description{
Run basic checks on curl get/post parameters
}
\keyword{internal}

