% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fb_copy_ad.R
\name{fbad_copy_ad}
\alias{fbad_copy_ad}
\title{Create a copy of an ad.}
\usage{
fbad_copy_ad(
  fbacc,
  ad_id,
  adset_id = NULL,
  status_option = NULL,
  rename_strategy = NULL,
  rename_prefix = NULL,
  rename_suffix = NULL,
  ...
)
}
\arguments{
\item{fbacc}{(optional) \code{FB_Ad_account} object, which defaults to the last returned object of \code{\link{fbad_init}}.}

\item{ad_id}{(numeric string or integer) id of the ad that you want to copy}

\item{adset_id}{(numeric string or integer) id adset you want to create a copy of the ad into}

\item{status_option}{(string) enum {ACTIVE, PAUSED, INHERITED_FROM_SOURCE}}

\item{rename_strategy}{(string) enum {DEEP_RENAME, ONLY_TOP_LEVEL_RENAME, NO_RENAME}}

\item{rename_prefix}{(string) a prefix to copy names. Defaults to null if not provided}

\item{rename_suffix}{(string) a suffix to copy names. Defaults to null if not provided}

\item{...}{further arguments passed to the API endpoint}
}
\description{
Create a copy of an ad.
}
\references{
\url{https://developers.facebook.com/docs/marketing-api/reference/adgroup/copies#Creating}
}
