% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fb_account.R
\name{fbad_create_account}
\alias{fbad_create_account}
\title{Create Ad Account}
\usage{
fbad_create_account(business_id, access_token, name, currency, timezone_id,
  end_advertiser = "NONE", media_agency = "NONE", partner = "NONE", ...)
}
\arguments{
\item{business_id}{Business Manager id}

\item{access_token}{token}

\item{name}{string}

\item{currency}{ISO 4217 Currency Code}

\item{timezone_id}{\url{https://developers.facebook.com/docs/marketing-api/reference/ad-account/timezone-ids}}

\item{end_advertiser}{string}

\item{media_agency}{string}

\item{partner}{string}

\item{...}{further parameters passed to the Facebook API}
}
\value{
Ad Account id
}
\description{
Create Ad Account
}
\references{
\url{https://developers.facebook.com/docs/marketing-api/reference/ad-account#Creating}
}
