% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/fb_adset.R
\name{fbad_create_adset}
\alias{fbad_create_adset}
\title{Create Ad Set}
\usage{
fbad_create_adset(fbacc, name, optimization_goal = c("NONE", "APP_INSTALLS",
  "CLICKS", "ENGAGED_USERS", "EXTERNAL", "EVENT_RESPONSES", "IMPRESSIONS",
  "LINK_CLICKS", "OFFER_CLAIMS", "OFFSITE_CONVERSIONS", "PAGE_ENGAGEMENT",
  "PAGE_LIKES", "POST_ENGAGEMENT", "REACH", "SOCIAL_IMPRESSIONS",
  "VIDEO_VIEWS"), billing_event = c("APP_INSTALLS", "CLICKS", "IMPRESSIONS",
  "LINK_CLICKS", "NONE", "OFFER_CLAIMS", "PAGE_LIKES", "POST_ENGAGEMENT"),
  bid_amount, bid_type = c("CPC", "CPM", "ABSOLUTE_OCPM", "CPA"), bid_info,
  promoted_object, campaign_group_id, campaign_status = c("ACTIVE", "PAUSED",
  "ARCHIVED", "DELETED"), daily_budget, lifetime_budget, end_time, start_time,
  targeting, ...)
}
\arguments{
\item{fbacc}{(optional) \code{FB_Ad_account} object, which defaults to the last returned object of \code{\link{fbad_init}}.}

\item{name}{name of the Ad Set}

\item{optimization_goal}{v2.4 only parameter}

\item{billing_event}{v2.4 only parameter}

\item{bid_amount}{v2.4 only parameter}

\item{bid_type}{v2.3 only parameter}

\item{bid_info}{v2.3 only parameter}

\item{promoted_object}{see at \url{https://developers.facebook.com/docs/marketing-api/reference/ad-campaign/promoted-object/v2.4}}

\item{campaign_group_id}{parent Ad Campaign id}

\item{campaign_status}{Ad Set status}

\item{daily_budget}{in account currency}

\item{lifetime_budget}{in account currency}

\item{end_time}{datetime}

\item{start_time}{datetime}

\item{targeting}{list}

\item{...}{further arguments passed to the API endpoint}
}
\value{
Ad Set id
}
\description{
Create Ad Set
}
\references{
\url{https://developers.facebook.com/docs/marketing-api/adset/v2.4#create}
}

