% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/fb.R
\name{fbad_create_audiance}
\alias{fbad_create_audiance}
\title{Create a new FB custom audiance}
\usage{
fbad_create_audiance(fbacc, name, description, opt_out_link)
}
\arguments{
\item{fbacc}{FB_Ad_account object returned by \code{fbad_init}}

\item{name}{string}

\item{description}{optional string}

\item{opt_out_link}{optional link}
}
\value{
custom audience ID
}
\description{
Create a new FB custom audiance
}
\references{
https://developers.facebook.com/docs/marketing-api/custom-audience-targeting/v2.2#create
}

