% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fb_audience.R
\name{fbad_create_audience}
\alias{fbad_create_audience}
\title{Create a new FB custom audience}
\usage{
fbad_create_audience(
  fbacc,
  name,
  description,
  subtype = c("CUSTOM", "WEBSITE", "APP", "OFFLINE_CONVERSION", "CLAIM", "PARTNER",
    "MANAGED", "VIDEO", "LOOKALIKE", "ENGAGEMENT", "DATA_SET", "BAG_OF_ACCOUNTS"),
  ...
)
}
\arguments{
\item{fbacc}{(optional) \code{FB_Ad_account} object, which defaults to the last returned object of \code{\link{fbad_init}}.}

\item{name}{string}

\item{description}{optional string}

\item{subtype}{audience type}

\item{...}{any further parameters (fields) passed to the API}
}
\value{
custom audience ID
}
\description{
Create a new FB custom audience
}
\references{
\url{https://developers.facebook.com/docs/marketing-api/reference/custom-audience#Creating}
}
