% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/fb_audience.R
\name{fbad_delete_audience}
\alias{fbad_delete_audience}
\title{Delete a FB custom audience}
\usage{
fbad_delete_audience(fbacc, audience_id)
}
\arguments{
\item{fbacc}{FB_Ad_account object returned by \code{fbad_init}}

\item{audience_id}{numeric}
}
\value{
custom audience ID
}
\description{
Delete a FB custom audience
}
\references{
\url{https://developers.facebook.com/docs/marketing-api/custom-audience-targeting/v2.3#delete}
}

