% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/fb_init.R
\name{fbad_get_baseurl}
\alias{fbad_get_baseurl}
\title{Get versioned base url}
\usage{
fbad_get_baseurl()
}
\arguments{
\item{version}{the version for which a base url is being generated}
}
\value{
character URL with trailing slash
}
\description{
Get versioned base url
}
\keyword{internal}

