% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fb_init.R
\name{fbad_get_client_ad_accounts}
\alias{fbad_get_client_ad_accounts}
\title{Deprecated in favor of \code{fbad_get_owned_adaccounts}}
\usage{
fbad_get_client_ad_accounts(id, token, version, fields = c("name"),
  simplify = TRUE)
}
\arguments{
\item{id}{Facebook Object, eg Ad Account (with \code{act} prefix) or a Business Manager Account ID}

\item{token}{FB Ads API token (if running before \code{fb_init})}

\item{version}{Facebook Marketing API version (if running before \code{fb_init})}

\item{fields}{character vector}

\item{simplify}{return \code{data.frame} or \code{list}}
}
\description{
Deprecated in favor of \code{fbad_get_owned_adaccounts}
}
