% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fb_audience.R
\name{fbad_get_customaudiences}
\alias{fbad_get_customaudiences}
\title{Obtain list of all existing custom audience}
\usage{
fbad_get_customaudiences(fbacc, audience_id, fields = c("id", "account_id",
  "approximate_count", "data_source", "delivery_status",
  "lookalike_audience_ids", "lookalike_spec", "name", "permission_for_actions",
  "operation_status", "subtype", "time_updated"))
}
\arguments{
\item{fbacc}{(optional) \code{FB_Ad_account} object, which defaults to the last returned object of \code{\link{fbad_init}}.}

\item{name}{string}

\item{origin_audience_id}{numeric ID of origin custom audience}

\item{ratio}{Between 0.01-0.20 and increments of 0.01. Indicates the top \code{ratio} percent of original audience in the selected country}

\item{country}{Country name - the country to find the lookalike people.}
}
\value{
lookalike audience ID and name
}
\description{
Obtain list of all existing custom audience
}
\references{
\url{https://developers.facebook.com/docs/marketing-api/lookalike-audience-targeting#create}
}

