% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fb_init.R
\name{fbad_get_my_ad_accounts}
\alias{fbad_get_my_ad_accounts}
\title{Get account details of Ad Accounts that are accessible by the given token}
\usage{
fbad_get_my_ad_accounts(token, version)
}
\arguments{
\item{token}{FB Ads API token (if running before \code{fb_init})}

\item{version}{Facebook Marketing API version (if running before \code{fb_init})}
}
\value{
character vector of Ad Account ids
}
\description{
Get account details of Ad Accounts that are accessible by the given token
}
