% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fb_init.R
\name{fbad_get_pixels}
\alias{fbad_get_pixels}
\title{Get tracking pixels of eg an Ad or Business Manager Account}
\usage{
fbad_get_pixels(id, token, version, fields = c("name"), simplify = TRUE)
}
\arguments{
\item{id}{Facebook Object, eg Ad Account (with \code{act} prefix) or a Business Manager Account ID}

\item{token}{FB Ads API token}

\item{version}{Facebook Marketing API version}

\item{fields}{character vector}

\item{simplify}{return \code{data.frame} or \code{list}}
}
\value{
list(s) containing Ads Pixels
}
\description{
Get tracking pixels of eg an Ad or Business Manager Account
}
\references{
\url{https://developers.facebook.com/docs/marketing-api/reference/ads-pixel/#Reading}
}
