% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fb_search.R
\name{fbad_get_search}
\alias{fbad_get_search}
\title{FB Search API Querying}
\usage{
fbad_get_search(fbacc, q, type = c("adeducationschool",
  "adeducationmajor", "adgeolocation", "adcountry", "adzipcode",
  "adgeolocationmeta", "adradiussuggestion", "adinterest",
  "adinterestsuggestion", "adinterestvalid", "adlocale",
  "adTargetingCategory", "adworkemployer", "targetingsearch"), ...)
}
\arguments{
\item{fbacc}{(optional) \code{FB_Ad_account} object, which defaults to the last returned object of \code{\link{fbad_init}}.}

\item{q}{string that is being searched for}

\item{type}{describes the type of search eg: adinterest, adeducationmajor etc}

\item{...}{other optional parameters accepted by the endpoint as key = value pairs eg: \code{limit = 5000}.}
}
\value{
\code{data.frame} containing results
}
\description{
FB Search API Querying
}
\examples{
\dontrun{
fbad_get_search(q = 'r programming language', type = 'adinterest')
fbad_get_search(q = c('dog', 'cat'), type = 'adinterestvalid')
}
}
\references{
\url{https://developers.facebook.com/docs/marketing-api/targeting-search}
}
