% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fb_insights.R
\name{fbad_insights_get_async_results}
\alias{fbad_insights_get_async_results}
\title{Wait for and get asynchronous report results}
\usage{
fbad_insights_get_async_results(fbacc, id, original_call, original_env)
}
\arguments{
\item{fbacc}{(optional) \code{FB_Ad_account} object, which defaults to the last returned object of \code{\link{fbad_init}}.}

\item{id}{job ID}

\item{original_call}{original call of \code{fb_insights} for future retries}

\item{original_env}{original environment of \code{fb_insights}}
}
\value{
JSON
}
\description{
Wait for and get asynchronous report results
}
\keyword{internal}

