% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fb_audience.R
\name{fbad_list_audience}
\alias{fbad_list_audience}
\title{List all Custom Audiences for Ad account}
\usage{
fbad_list_audience(fbacc, id, statuses, fields = "id", simplify = TRUE)
}
\arguments{
\item{fbacc}{(optional) \code{FB_Ad_account} object, which defaults to the last returned object of \code{\link{fbad_init}}.}

\item{id}{will do the look-up for all Ads based on this ID. Defaults to current FB account. Can be a (vector of) Ad Set or Campaign id(s).}

\item{statuses}{character vector of statuses to use as a filter. Defaults to none. Please refer to the Facebook documentation for a list of possible values.}

\item{fields}{character vector of fields to get from the API, defaults to \code{id}. Please refer to the Facebook documentation for a list of possible values.}

\item{simplify}{boolean whether response is simplified to data.frame or else returned as raw list}
}
\description{
List all Custom Audiences for Ad account
}
\references{
\url{https://developers.facebook.com/docs/marketing-api/reference/ad-account/customaudiences/#Reading}
}
