% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/fb_campaign.R
\name{fbad_list_campaign}
\alias{fbad_list_campaign}
\title{List all Ad Campaigns for current account}
\usage{
fbad_list_campaign(fbacc, statuses, fields = "id")
}
\arguments{
\item{fbacc}{(optional) \code{FB_Ad_account} object, which defaults to the last returned object of \code{\link{fbad_init}}.}

\item{statuses}{character vector of statuses to use as a filter. Defaults to none. Please refer to the Facebook documentation for a list of possible values.}

\item{fields}{character vector of fields to get from the API, defaults to \code{id}. Please refer to the Facebook documentation for a list of possible values.}
}
\description{
List all Ad Campaigns for current account
}
\references{
\url{https://developers.facebook.com/docs/marketing-api/reference/ad-campaign-group#Reading}
}

