% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/fb_ad.R
\name{fbad_preview_ad}
\alias{fbad_preview_ad}
\title{Preview ad}
\usage{
fbad_preview_ad(fbacc, id, ad_format = c("DESKTOP_FEED_STANDARD",
  "RIGHT_COLUMN_STANDARD", "MOBILE_FEED_STANDARD", "MOBILE_BANNER",
  "MOBILE_INTERSTITIAL", "INSTAGRAM_STANDARD"))
}
\arguments{
\item{fbacc}{(optional) \code{FB_Ad_account} object, which defaults to the last returned object of \code{\link{fbad_init}}.}

\item{id}{ad id(s)}
}
\description{
Preview ad
}
\references{
\url{https://developers.facebook.com/docs/marketing-api/generatepreview}
}

