% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/fb_ad.R
\name{fbad_read_ad}
\alias{fbad_read_ad}
\title{Read ad}
\usage{
fbad_read_ad(fbacc, id, fields = "id")
}
\arguments{
\item{fbacc}{(optional) \code{FB_Ad_account} object, which defaults to the last returned object of \code{\link{fbad_init}}.}

\item{id}{ad id(s)}

\item{fields}{character vector of fields to get from the API -- please refer to the FB documentation for a list of possible values.}
}
\value{
data.frame
}
\description{
Read ad
}
\references{
\url{https://developers.facebook.com/docs/marketing-api/adgroup/v2.4#read}
}

