% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fb_adset.R
\name{fbad_read_adset}
\alias{fbad_read_adset}
\title{Read Ad Set details}
\usage{
fbad_read_adset(fbacc, id, fields = "id")
}
\arguments{
\item{fbacc}{(optional) \code{FB_Ad_account} object, which defaults to the last returned object of \code{\link{fbad_init}}.}

\item{id}{ad set id(s)}

\item{fields}{character vector of fields to get from the API, defaults to \code{id}. Please refer to the Facebook documentation for a list of possible values.}
}
\description{
Read Ad Set details
}
\references{
\url{https://developers.facebook.com/docs/marketing-api/reference/ad-campaign#Reading}
}

