% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/fb_campaign.R
\name{fbad_read_campaign}
\alias{fbad_read_campaign}
\title{Read ad campaign}
\usage{
fbad_read_campaign(fbacc, id, fields = c("id", "account_id", "adgroups",
  "buying_type", "campaign_group_status", "objective", "spend_cap", "name"))
}
\arguments{
\item{fbacc}{(optional) \code{FB_Ad_account} object, which defaults to the last returned object of \code{\link{fbad_init}}.}

\item{id}{}

\item{fields}{}
}
\value{
list
}
\description{
Read ad campaign
}
\references{
\url{https://developers.facebook.com/docs/marketing-api/adcampaign/v2.3$read}
}

