% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/fb_audience.R
\name{fbad_remove_audience}
\alias{fbad_remove_audience}
\title{Add people from a custom FB audience}
\usage{
fbad_remove_audience(fbacc, audience_id, schema = c("EMAIL", "PHONE"), hashes)
}
\arguments{
\item{fbacc}{(optional) \code{FB_Ad_account} object, which defaults to the last returned object of \code{\link{fbad_init}}.}

\item{audience_id}{string}

\item{schema}{only two schema are supported out of the four: you can add/remove persons to/from a custom audience by e-mail addresses or phone numbers}

\item{hashes}{character vector of e-mail addresses or phone numbers to be transformed to hashes}
}
\description{
Add people from a custom FB audience
}
\references{
\url{https://developers.facebook.com/docs/marketing-api/reference/custom-audience/users#Deleting}
}

