% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fb_init.R
\name{fbad_request_next_page}
\alias{fbad_request_next_page}
\title{Fetch the content of the URL and return JSON}
\usage{
fbad_request_next_page(url)
}
\arguments{
\item{url}{next URL as returned by FB}
}
\description{
Fetch the content of the URL and return JSON
}
\keyword{internal}
