% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fb_audience.R
\name{fbad_share_audience}
\alias{fbad_share_audience}
\title{Share a FB custom audience with other accounts}
\usage{
fbad_share_audience(fbacc, audience_id, adaccounts)
}
\arguments{
\item{fbacc}{(optional) \code{FB_Ad_account} object, which defaults to the last returned object of \code{\link{fbad_init}}.}

\item{audience_id}{audience ID}

\item{adaccounts}{numeric vector of FB account IDs}
}
\description{
Share a FB custom audience with other accounts
}
\note{
This throws error if you provide wrong account ids OR even valid account ids that were previously granted access to the given custom audience.
}
