% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/fb_ad.R
\name{fbad_update_ad}
\alias{fbad_update_ad}
\title{Update ad}
\usage{
fbad_update_ad(fbacc, id, ...)
}
\arguments{
\item{fbacc}{(optional) \code{FB_Ad_account} object, which defaults to the last returned object of \code{\link{fbad_init}}.}

\item{id}{ad id}

\item{...}{parameters passed to Facebook API}
}
\value{
invisible TRUE
}
\description{
Update ad
}
\references{
\url{https://developers.facebook.com/docs/marketing-api/adgroup/v2.4#update}
}

