% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/fb_adset.R
\name{fbad_update_adset}
\alias{fbad_update_adset}
\title{Update adset}
\usage{
fbad_update_adset(fbacc, id, ...)
}
\arguments{
\item{fbacc}{(optional) \code{FB_Ad_account} object, which defaults to the last returned object of \code{\link{fbad_init}}.}

\item{id}{adset id}

\item{...}{parameters passed to Facebook API}
}
\description{
Update adset
}
\references{
\url{https://developers.facebook.com/docs/marketing-api/reference/ad-campaign#Updating}
}

