% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fb_init.R
\name{fbad_whoami}
\alias{fbad_whoami}
\title{Prints user id and name}
\usage{
fbad_whoami(token, version)
}
\arguments{
\item{token}{FB Ads API token (if running before \code{fb_init})}

\item{version}{Facebook Marketing API version (if running before \code{fb_init})}
}
\value{
character vector of Ad Account ids
}
\description{
Prints user id and name
}
