% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zzz.R
\docType{package}
\name{fbRads}
\alias{fbRads}
\alias{fbRads-package}
\title{fbRads package}
\description{
This is a placeholder for storing import directives, please find more details in the \code{README.md} file of the package via \code{system.file} or on GitHub at \url{https://github.com/cardcorp/fbRads}.
}
