% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{fromJSONish}
\alias{fromJSONish}
\title{Validates and fixes some JSON issues, eg removing newlines etc}
\usage{
fromJSONish(json, ...)
}
\arguments{
\item{json}{string}

\item{...}{passed to jsonlite}
}
\value{
parsed JSON as an R object
}
\description{
Validates and fixes some JSON issues, eg removing newlines etc
}
