% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fb_init.R
\name{is.FB_Ad_Account}
\alias{is.FB_Ad_Account}
\title{Checks if object is a valid initialized connection to the Facebook Marketing API}
\usage{
is.FB_Ad_Account(fbacc)
}
\arguments{
\item{fbacc}{R object with \code{FB_Ad_Account} class}
}
\description{
Checks if object is a valid initialized connection to the Facebook Marketing API
}
\keyword{internal}
