% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fb_init.R
\name{print.FB_Ad_Account}
\alias{print.FB_Ad_Account}
\title{Print method for custom fbRads class}
\usage{
\method{print}{FB_Ad_Account}(x, ...)
}
\arguments{
\item{x}{R object with \code{FB_Ad_Account} class}

\item{...}{further arguments passed to \code{print} (ignored)}
}
\description{
Print method for custom fbRads class
}
