% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{this_function_name}
\alias{this_function_name}
\title{Return the name of the parent function}
\usage{
this_function_name()
}
\value{
string
}
\description{
Return the name of the parent function
}
\keyword{internal}
