% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{url_parse}
\alias{url_parse}
\title{Parse Facebook URL without bringing in another dependency. Might change this for urltools or httr}
\usage{
url_parse(url)
}
\arguments{
\item{url}{string}
}
\value{
list
}
\description{
Parse Facebook URL without bringing in another dependency. Might change this for urltools or httr
}
\keyword{internal}
