% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/fb_audience.R
\name{fbad_add_audience}
\alias{fbad_add_audience}
\title{FB add people to audience}
\usage{
fbad_add_audience(fbacc, audience_id, schema = c("EMAIL", "PHONE"), hashes)
}
\arguments{
\item{fbacc}{(optional) \code{FB_Ad_account} object, which defaults to the last returned object of \code{\link{fbad_init}}.}

\item{audience_id}{string}

\item{schema}{only two schema are supported out of the four: you can add persons to a custom audience by e-mail addresses or phone numbers}

\item{hashes}{character vector of e-mail addresses or phone numbers to be transformed to hashes}
}
\description{
FB add people to audience
}
\references{
\url{https://developers.facebook.com/docs/marketing-api/custom-audience-targeting/v2.3#create}
}

