% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/fb_init.R
\name{fbad_check_fbacc}
\alias{fbad_check_fbacc}
\title{Check if provided R object is a valid list of FB account details}
\usage{
fbad_check_fbacc(fbacc)
}
\arguments{
\item{fbacc}{R object}
}
\description{
Check if provided R object is a valid list of FB account details
}
\keyword{intenral}

