% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/fb_init.R
\name{fbad_check_fbacc}
\alias{fbad_check_fbacc}
\title{Check if provided or previously initialized  R object is a valid FB Ad Account}
\usage{
fbad_check_fbacc(fbacc)
}
\arguments{
\item{fbacc}{(optional) \code{FB_Ad_account} object, which defaults to the last returned object of \code{\link{fbad_init}}.}
}
\value{
invisibly returned \code{FB_Ad_Account} object
}
\description{
Check if provided or previously initialized  R object is a valid FB Ad Account
}
\keyword{internal}

