% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/fb_ad.R
\name{fbad_create_ad}
\alias{fbad_create_ad}
\title{Create ad}
\usage{
fbad_create_ad(fbacc, name, campaign_id, creative_id,
  adgroup_status = c("ACTIVE", "PAUSED"))
}
\arguments{
\item{fbacc}{}

\item{name}{}

\item{campaign_id}{}

\item{creative_id}{}

\item{adgroup_status}{}
}
\value{
ad id
}
\description{
Create ad
}
\references{
\url{https://developers.facebook.com/docs/marketing-api/adgroup/v2.3#create}
}

