% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fb_ad.R
\name{fbad_create_ad}
\alias{fbad_create_ad}
\title{Create Ad}
\usage{
fbad_create_ad(fbacc, ...)
}
\arguments{
\item{fbacc}{(optional) \code{FB_Ad_account} object, which defaults to the last returned object of \code{\link{fbad_init}}.}

\item{...}{further parameters passed to the API, see below references}
}
\value{
ad id
}
\description{
Create Ad
}
\references{
\url{https://developers.facebook.com/docs/marketing-api/reference/adgroup#Creating}
}
