% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/fb_ad.R
\name{fbad_create_ad}
\alias{fbad_create_ad}
\title{Create ad}
\usage{
fbad_create_ad(fbacc, name, campaign_id, creative_id,
  adgroup_status = c("ACTIVE", "PAUSED"), ...)
}
\arguments{
\item{fbacc}{(optional) \code{FB_Ad_account} object, which defaults to the last returned object of \code{\link{fbad_init}}.}

\item{name}{Ad group name}

\item{campaign_id}{Ad Set id}

\item{creative_id}{creative ID}

\item{adgroup_status}{initial status of the Ad group}

\item{...}{further parameters passed to the Facebook API}
}
\value{
ad id
}
\description{
Create ad
}
\references{
\url{https://developers.facebook.com/docs/marketing-api/adgroup/v2.4#create}
}

