% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/fb_adset.R
\name{fbad_create_adset}
\alias{fbad_create_adset}
\title{Create adset}
\usage{
fbad_create_adset(fbacc, name, optimization_goal = c("NONE", "APP_INSTALLS",
  "CLICKS", "ENGAGED_USERS", "EXTERNAL", "EVENT_RESPONSES", "IMPRESSIONS",
  "LINK_CLICKS", "OFFER_CLAIMS", "OFFSITE_CONVERSIONS", "PAGE_ENGAGEMENT",
  "PAGE_LIKES", "POST_ENGAGEMENT", "REACH", "SOCIAL_IMPRESSIONS",
  "VIDEO_VIEWS"), billing_event = c("APP_INSTALLS", "CLICKS", "IMPRESSIONS",
  "LINK_CLICKS", "NONE", "OFFER_CLAIMS", "PAGE_LIKES", "POST_ENGAGEMENT"),
  bid_amount, bid_type = c("CPC", "CPM", "ABSOLUTE_OCPM", "CPA"), bid_info,
  campaign_group_id, campaign_status = c("ACTIVE", "PAUSED", "ARCHIVED",
  "DELETED"), daily_budget, lifetime_budget, end_time, start_time, targeting,
  ...)
}
\arguments{
\item{fbacc}{(optional) \code{FB_Ad_account} object, which defaults to the last returned object of \code{\link{fbad_init}}.}

\item{name}{}

\item{optimization_goal}{v2.4 only parameter}

\item{billing_event}{v2.4 only parameter}

\item{bid_amount}{v2.4 only parameter}

\item{bid_type}{v2.3 only parameter}

\item{bid_info}{v2.3 only parameter}

\item{campaign_status}{}

\item{daily_budget}{}

\item{lifetime_budget}{}

\item{end_time}{}

\item{start_time}{}

\item{targeting}{}

\item{...}{}
}
\value{
adset id
}
\description{
Create adset
}
\references{
\url{https://developers.facebook.com/docs/marketing-api/adset/v2.4#create}
}

