% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fb_adset.R
\name{fbad_create_adset}
\alias{fbad_create_adset}
\title{Create Ad Set}
\usage{

}
\arguments{
\item{fbacc}{(optional) \code{FB_Ad_account} object, which defaults to the last returned object of \code{\link{fbad_init}}.}

\item{...}{further parameters passed to the API, see below references}
}
\value{
Ad Set id
}
\description{
Create Ad Set
}
\references{
\url{https://developers.facebook.com/docs/marketing-api/reference/ad-campaign#Creating}
}
