% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/fb_adset.R
\name{fbad_create_adset}
\alias{fbad_create_adset}
\title{Create adset}
\usage{
fbad_create_adset(fbacc, name, bid_type = c("CPC", "CPM", "ABSOLUTE_OCPM",
  "CPA"), bid_info, campaign_group_id, campaign_status = c("ACTIVE", "PAUSED",
  "ARCHIVED", "DELETED"), daily_budget, lifetime_budget, end_time, start_time,
  targeting)
}
\arguments{
\item{fbacc}{(optional) \code{FB_Ad_account} object, which defaults to the last returned object of \code{\link{fbad_init}}.}

\item{name}{}

\item{bid_type}{}

\item{bid_info}{}

\item{campaign_group_id}{}

\item{campaign_status}{}

\item{daily_budget}{}

\item{lifetime_budget}{}

\item{end_time}{}

\item{start_time}{}

\item{targeting}{}
}
\value{
adset id
}
\description{
Create adset
}
\references{
\url{https://developers.facebook.com/docs/marketing-api/adset/v2.3#create}
}

