% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/fb_audience.R
\name{fbad_create_audience}
\alias{fbad_create_audience}
\title{Create a new FB custom audience}
\usage{
fbad_create_audience(fbacc, name, description, opt_out_link)
}
\arguments{
\item{fbacc}{(optional) \code{FB_Ad_account} object, which defaults to the last returned object of \code{\link{fbad_init}}.}

\item{name}{string}

\item{description}{optional string}

\item{opt_out_link}{optional link}
}
\value{
custom audience ID
}
\description{
Create a new FB custom audience
}
\references{
\url{https://developers.facebook.com/docs/marketing-api/custom-audience-targeting/v2.4#create}
}

