% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fb_campaign.R
\name{fbad_create_campaign}
\alias{fbad_create_campaign}
\title{Create Ad Campaign}
\usage{
fbad_create_campaign(fbacc, ...)
}
\arguments{
\item{fbacc}{(optional) \code{FB_Ad_account} object, which defaults to the last returned object of \code{\link{fbad_init}}.}

\item{...}{further parameters passed to the API, see below references}
}
\value{
Ad Campaign id
}
\description{
Create Ad Campaign
}
\references{
\url{https://developers.facebook.com/docs/marketing-api/reference/ad-campaign-group#Creating}
}
