% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/fb_campaign.R
\name{fbad_create_campaign}
\alias{fbad_create_campaign}
\title{Created ad campaign}
\usage{
fbad_create_campaign(fbacc, buying_type = c("AUCTION", "FIXED_CPM",
  "RESERVED"), campaign_group_status = c("ACTIVE", "PAUSED"),
  execution_options = NULL, name, objective = c("NONE",
  "CANVAS_APP_ENGAGEMENT", "CANVAS_APP_INSTALLS", "EVENT_RESPONSES",
  "LOCAL_AWARENESS", "MOBILE_APP_ENGAGEMENT", "MOBILE_APP_INSTALLS",
  "OFFER_CLAIMS", "PAGE_LIKES", "POST_ENGAGEMENT", "VIDEO_VIEWS",
  "WEBSITE_CLICKS", "WEBSITE_CONVERSIONS"), spend_cap = NULL)
}
\arguments{
\item{fbacc}{(optional) \code{FB_Ad_account} object, which defaults to the last returned object of \code{\link{fbad_init}}.}

\item{buying_type}{}

\item{campaign_group_status}{}

\item{execution_options}{}

\item{name}{}

\item{objective}{}

\item{spend_cap}{}
}
\value{
ad campaign id
}
\description{
Created ad campaign
}
\references{
\url{https://developers.facebook.com/docs/marketing-api/adcampaign/v2.4#create}
}

