% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/fb_creative.R
\name{fbad_create_creative}
\alias{fbad_create_creative}
\title{Create an ad creative}
\usage{
fbad_create_creative(fbacc, title, body, name, actor_id, follow_redirect,
  image_file, image_hash, image_crops, image_url, link_url, url_tags, object_id,
  object_story_id, object_story_spec, object_url,
  call_to_action_type = c("OPEN_LINK", "BOOK_TRAVEL", "SHOP_NOW", "PLAY_GAME",
  "LISTEN_MUSIC", "WATCH_VIDEO", "USE_APP"))
}
\arguments{
\item{fbacc}{(optional) \code{FB_Ad_account} object, which defaults to the last returned object of \code{\link{fbad_init}}.}

\item{title}{}

\item{body}{}

\item{name}{}

\item{actor_id}{}

\item{follow_redirect}{}

\item{image_file}{}

\item{image_hash}{}

\item{image_crops}{}

\item{image_url}{}

\item{link_url}{}

\item{url_tags}{}

\item{object_id}{}

\item{object_story_id}{}

\item{object_story_spec}{}

\item{object_url}{}

\item{call_to_action_type}{}
}
\value{
creative id
}
\description{
Create an ad creative
}
\note{
There are no checks done before passing provided arguments to Facebook. You have to know what you are up to. Read the Facebook docs.
}
\references{
\url{https://developers.facebook.com/docs/marketing-api/adcreative/v2.3#create}
}

