% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/fb_image.R
\name{fbad_create_image}
\alias{fbad_create_image}
\title{Upload image}
\usage{
fbad_create_image(fbacc, img)
}
\arguments{
\item{fbacc}{(optional) \code{FB_Ad_account} object, which defaults to the last returned object of \code{\link{fbad_init}}.}

\item{img}{}
}
\value{
list of file name, hash and URL
}
\description{
Upload image
}
\references{
\url{https://developers.facebook.com/docs/marketing-api/adimage/v2.4#create}
}

