% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fb_audience.R
\name{fbad_create_lookalike_audience}
\alias{fbad_create_lookalike_audience}
\title{Create a new FB lookalike audience similar to an already existing custom audience}
\usage{
fbad_create_lookalike_audience(
  fbacc,
  name,
  origin_audience_id,
  ratio = 0.01,
  country = "US"
)
}
\arguments{
\item{fbacc}{(optional) \code{FB_Ad_account} object, which defaults to the last returned object of \code{\link{fbad_init}}.}

\item{name}{string}

\item{origin_audience_id}{numeric ID of origin custom audience}

\item{ratio}{Between 0.01-0.20 and increments of 0.01. Indicates the top \code{ratio} percent of original audience in the selected country}

\item{country}{Country name - the country to find the lookalike people.}
}
\value{
lookalike audience ID
}
\description{
Create a new FB lookalike audience similar to an already existing custom audience
}
\references{
\url{https://developers.facebook.com/docs/marketing-api/lookalike-audience-targeting#create}
}
