% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/fb_init.R
\name{fbad_get_adaccount_details}
\alias{fbad_get_adaccount_details}
\title{Get details for a Facebook Ads Account}
\usage{
fbad_get_adaccount_details(accountid, token, version)
}
\arguments{
\item{accountid}{Ads account graph object id}

\item{token}{FB Ads API token}

\item{version}{Facebook Marketing API version}
}
\value{
list(s) containing account details
}
\description{
Get details for a Facebook Ads Account
}
\references{
\url{https://developers.facebook.com/docs/marketing-api/reference/ad-account}
}

