% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fb_init.R
\name{fbad_get_adaccounts}
\alias{fbad_get_adaccounts}
\title{Get account details belonging to eg a Business Manager Account}
\usage{
fbad_get_adaccounts(id, token, version, fields = c("name"))
}
\arguments{
\item{id}{Facebook Object, eg a Business Manager Account ID}

\item{token}{FB Ads API token}

\item{version}{Facebook Marketing API version}

\item{fields}{character vector}
}
\value{
list(s) containing account details
}
\description{
Get account details belonging to eg a Business Manager Account
}
\references{
\url{https://developers.facebook.com/docs/marketing-api/reference/ad-account#Reading}
}

