% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/fb_insights.R
\name{fbad_insights_get_async_results}
\alias{fbad_insights_get_async_results}
\title{Wait for and get asynchronous report results}
\usage{
fbad_insights_get_async_results(id)
}
\arguments{
\item{id}{job ID}
}
\value{
JSON
}
\description{
Wait for and get asynchronous report results
}
\keyword{internal}

