% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/fb_ad.R
\name{fbad_list_ad}
\alias{fbad_list_ad}
\title{List ads for current account}
\usage{
fbad_list_ad(fbacc, fields = "id")
}
\arguments{
\item{fbacc}{(optional) \code{FB_Ad_account} object, which defaults to the last returned object of \code{\link{fbad_init}}.}

\item{fields}{character vector of fields to get from the API, defaults to \code{id}. Please refer to the Facebook documentation for a list of possible values.}
}
\value{
data.frame
}
\description{
List ads for current account
}
\note{
Will do a batched request to the Facebook API if multiple ids are provided.
}
\references{
\url{https://developers.facebook.com/docs/marketing-api/adgroup/v2.4#read-adaccount}
}

