% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fb_audience.R
\name{fbad_list_audience}
\alias{fbad_list_audience}
\title{List all Custom Audiences for current Ad account}
\usage{
fbad_list_audience(fbacc, id, statuses, fields = "id")
}
\arguments{
\item{fbacc}{(optional) \code{FB_Ad_account} object, which defaults to the last returned object of \code{\link{fbad_init}}.}

\item{fields}{character vector of fields to get from the API, defaults to \code{id}. Please refer to the Facebook documentation for a list of possible values.}
}
\description{
List all Custom Audiences for current Ad account
}
\references{
\url{https://developers.facebook.com/docs/marketing-api/reference/ad-account/customaudiences/#Reading}
}
