% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/fb.R
\name{fbad_reachestimate}
\alias{fbad_reachestimate}
\title{Query for reach estimate for given targeting spec}
\usage{
fbad_reachestimate(fbacc, targeting_spec, currency = "USD")
}
\arguments{
\item{fbacc}{FB_Ad_account object returned by \code{fbad_init}}

\item{targeting_spec}{lists of targeting spec characteristics}

\item{currency}{string}
}
\value{
list
}
\description{
Query for reach estimate for given targeting spec
}
\examples{
\dontrun{
targetspec <- list(
  age_min = unbox(24),
  age_max = unbox(55),
  geo_locations = list(countries = 'US'))
}
}

