% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/fb_campaign.R
\name{fbad_read_campaign}
\alias{fbad_read_campaign}
\title{Read Ad Campaign details}
\usage{
fbad_read_campaign(fbacc, id, fields = "id")
}
\arguments{
\item{fbacc}{(optional) \code{FB_Ad_account} object, which defaults to the last returned object of \code{\link{fbad_init}}.}

\item{id}{Ad Campaign id(s)}

\item{fields}{character vector of fields to get from the API, defaults to \code{id}. Please refer to the Facebook documentation for a list of possible values.}
}
\description{
Read Ad Campaign details
}
\references{
\url{https://developers.facebook.com/docs/marketing-api/adcampaign/v2.4$read}
}

