% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/fb_init.R
\name{fbad_request}
\alias{fbad_request}
\title{Get results of a synchronous query from FB graph API}
\usage{
fbad_request(path, method = c("GET", "POST", "DELETE"), params,
  debug = FALSE, log = TRUE)
}
\arguments{
\item{path}{API request path (i.e. endpoint)}

\item{method}{HTTP request type (e.g. GET or POST or DELETE)}

\item{params}{a name-value list of form parameters for API query}

\item{debug}{print debug messages by calling Curl verbosely}

\item{log}{print log messages or suppress those}
}
\value{
json object containing results
}
\description{
Get results of a synchronous query from FB graph API
}
\keyword{internal}

