% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fb_init.R
\name{fbad_request}
\alias{fbad_request}
\title{Get results of a synchronous query from FB graph API}
\usage{
fbad_request(fbacc, path, method = c("GET", "POST", "DELETE"),
  params = list(), debug = FALSE, log = TRUE, version, retries = 0)
}
\arguments{
\item{fbacc}{(optional) \code{FB_Ad_account} object, which defaults to the last returned object of \code{\link{fbad_init}}.}

\item{path}{API request path (i.e. endpoint)}

\item{method}{HTTP request type (e.g. GET or POST or DELETE)}

\item{params}{a name-value list of form parameters for API query}

\item{debug}{print debug messages by calling Curl verbosely}

\item{log}{print log messages or suppress those}

\item{version}{Facebook Marketing API version, defaults to what was used in the most recent \code{fbad_init} function call or \code{fb_api_most_recent_version()} before such function call}

\item{retries}{number of times the current query was tried previously -- used to handle network errors}
}
\value{
json object containing results
}
\description{
Get results of a synchronous query from FB graph API
}
\keyword{internal}
