% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/fb.R
\name{fbad_share_audience}
\alias{fbad_share_audience}
\title{Share a FB custom audience with other accounts}
\usage{
fbad_share_audience(fbacc, audience_id, adaccounts)
}
\arguments{
\item{fbacc}{FB_Ad_account object returned by \code{fbad_init}}

\item{audience_id}{audience ID}

\item{adaccounts}{numeric vector of FB account IDs}
}
\description{
Share a FB custom audience with other accounts
}
\note{
This throws error if you provide wrong account ids OR even valid account ids that were previously granted access to the given custom audience.
}
\references{
https://developers.facebook.com/docs/marketing-api/custom-audience-targeting/v2.2#sharing
}

