% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/zzz.R
\docType{package}
\name{fbRads}
\alias{fbRads}
\alias{fbRads-package}
\title{CARD.com internal R package}
\description{
This is a placeholder for storing import directives, please find more details in the \code{README.md} file.
}

